+++ HISTORY.md
@@ -5,6 +5,8 @@
 
 - Implemented a WebSocket transport.
 - Added support for `hypertimer`, for discrete event simulations.
+- Added support for a configurable `random` function , for deterministic random 
+  values.
 
 
 ## 2014-10-12, version 0.4.1